azure-function/requirements.txt
azure-functions==1.21.3
xgboost==2.1.2
scikit-learn==1.5.2
joblib==1.4.2
pandas==2.2.3
numpy==2.1.3
pyarrow==18.1.0
openai==1.55.0
